+++ packages/browser-pool/src/proxy-server.ts
@@ -10,7 +10,6 @@ const listenPromise = server.listen();
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 listenPromise.catch(() => {});
 
-// eslint-disable-next-line max-len
 // https://github.com/microsoft/playwright/blob/2e4722d460b5142267e0e506ca7ea9a259556b5f/packages/playwright-core/src/server/browserContext.ts#L423-L427
 export async function getLocalProxyAddress(): Promise<string> {
     await listenPromise;