+++ packages/browser-pool/src/utils.ts
@@ -3,7 +3,6 @@ import type { BrowserPlugin } from './abstract-classes/browser-plugin';
 
 export type UnwrapPromise<T> = T extends PromiseLike<infer R> ? UnwrapPromise<R> : T;
 
-// eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-unused-vars
 export function noop(..._args: unknown[]): void {}
 
 export type InferBrowserPluginArray<