+++ test/core/crawlers/basic_crawler.test.ts
@@ -129,7 +129,7 @@ describe.each(SingleStorageCase)('BasicCrawler - %s', (Emulator) => {
 
         // clean up
         // @ts-expect-error Accessing private method
-        await basicCrawler.autoscaledPool._destroy(); // eslint-disable-line no-underscore-dangle
+        await basicCrawler.autoscaledPool._destroy();
         getValueSpy.mockRestore();
         setValueSpy.mockRestore();
     });
@@ -372,7 +372,7 @@ describe.each(SingleStorageCase)('BasicCrawler - %s', (Emulator) => {
         });
 
         // @ts-expect-error Accessing private prop
-        expect(await crawler._isTaskReadyFunction()).toBe(false); // eslint-disable-line no-underscore-dangle
+        expect(await crawler._isTaskReadyFunction()).toBe(false);
     });
 
     test('should be possible to override isFinishedFunction of underlying AutoscaledPool', async () => {
@@ -839,7 +839,7 @@ describe.each(SingleStorageCase)('BasicCrawler - %s', (Emulator) => {
             });
 
             // @ts-expect-error Accessing private prop
-            crawler._loadHandledRequestCount = () => { // eslint-disable-line
+            crawler._loadHandledRequestCount = () => {
                 expect(crawler.sessionPool).toBeDefined();
                 expect(events.listenerCount(EventType.PERSIST_STATE)).toEqual(1);
             };