+++ test/core/crawlers/playwright_crawler.test.ts
@@ -4,7 +4,6 @@ import express from 'express';
 import playwright from 'playwright';
 import log from '@apify/log';
 import {
-    Configuration,
     PlaywrightCrawler,
     PlaywrightGotoOptions,
     PlaywrightRequestHandler,
@@ -81,12 +80,15 @@ describe.each(StorageTestCases)('PlaywrightCrawler - %s', (Emulator) => {
                 expect(response.status()).toBe(200);
                 request.userData.title = await page.title();
                 processed.push(request);
+                expect(response.request().headers()['user-agent']).not.toMatch(/headless/i);
+                await expect(page.evaluate(() => window.navigator.webdriver)).resolves.toBeFalsy();
             };
 
             const playwrightCrawler = new PlaywrightCrawler({
                 launchContext: {
                     launcher: playwright[browser],
                 },
+                browserPoolOptions: { useFingerprints: false },
                 requestList: requestListLarge,
                 minConcurrency: 1,
                 maxConcurrency: 1,