+++ test/core/request_list.test.ts
@@ -655,7 +655,7 @@ describe.each(StorageTestCases)('RequestList - %s', (Emulator) => {
         // Get requests from list
         let reqs: Request[] = [];
         for (let i = 0; i < 5; i++) {
-            const request = await requestList.fetchNextRequest(); // eslint-disable-line
+            const request = await requestList.fetchNextRequest();
             if (!request) break;
             reqs.push(request);
         }
@@ -663,7 +663,7 @@ describe.each(StorageTestCases)('RequestList - %s', (Emulator) => {
         reqs = shuffle(reqs) as typeof reqs;
 
         for (let i = 0; i < reqs.length; i++) {
-            await requestList.reclaimRequest(reqs[i]); // eslint-disable-line
+            await requestList.reclaimRequest(reqs[i]);
         }
     });